pyproject.toml
[project]
name = "ai-chatbot-backend"
version = "0.1.0"
description = "AI Chatbot with RAG, Auth, and History"
authors = [
    {name = "mdwasikhan54",email = "mdwasikhan54@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "sqlalchemy>=2.0.25",
    "pydantic[email]>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "langchain-community>=0.0.19",
    "openai>=1.12.0",
    "apscheduler>=3.10.4",
    "python-dotenv>=1.0.1",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "bcrypt (==3.2.0)"
]

[tool.poetry]
packages = [{include = "backend"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
